Move static composable and node to appropriate files
